Enabling the puppet3 future parser reveals changes needed for puppet4, specifically updates to class names which are no longer allowed to be numeric.
This is an initial update which fixes the classnames to support the
future parser.

This also updates hiera.yaml to rename the Ubuntu OS yaml lookup files
by prepending ${asfosname}_ , and removes the virtual and oem lookup
blocks as these are no longer in use.

This commit also contains an update to the python module to support
versioned pips. This requires explicit disabling of the gunicorn service.

Finally, this adds some additional packages for working on the migration
of the pootle service to a new VM.